+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes will be documented in this file.
 
 ## Head
 
+### Changed
+
+* Install Mambaforge (#351)
+
 ## 0.9.2 &ndash; 2023-02-22
 
 ### Added